requirements.txt
numpy
pandas
scikit-learn
matplotlib
plotly
tqdm
pyyaml
# DQN/PPO later:
stable-baselines3==2.3.2
torch --index-url https://download.pytorch.org/whl/cpu
